# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Government Template Platform V3.0 API",
    "version": "3.0.0",
    "description": "Цахим засаглалыг бүтээх суурь — eID нэвтрэлт, RBAC, AI туслах бүхий төрийн үйлчилгээний платформын API (Node.js · Express 5 · PostgreSQL · Redis)."
  },
  "servers": [
    {
      "url": "https://node.template.dgov.mn/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "BaseResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    },
    "parameters": {
      "IntegrationProvider": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "google-drive",
            "dropbox",
            "google-meet"
          ]
        }
      },
...
      "OrgRegNo": {
        "name": "regNo",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Байгууллагын улсын бүртгэлийн дугаар"
      },
...
    },
    "requestBodies": {
...
      "AssetURL": {
        "required": true,
        "description": "Google Drive-д байршуулсан зургийн URL",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "format": "uri",
                  "maxLength": 1000
                }
              },
              "required": [
                "url"
              ]
            }
          }
        }
      },
...
    },
    "responses": {
      "Ok": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BaseResponse"
            }
          }
        }
      },
      "Error": {
        "description": "Алдааны нэгдсэн дугтуй",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BaseResponse"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/integrations": {
      "get": {
        "summary": "Холбосон интеграцууд",
        "description": "Хэрэглэгчийн холбосон гуравдагч талын үйлчилгээнүүд. Хариунд ТОКЕН ОРОХГҮЙ — зөвхөн provider + хугацаа.",
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "summary": "Интеграцийн токеныг хадгалах",
        "description": "OAuth token exchange хийсний дараа токеныг backend-д AES-256-GCM-ээр ШИФРЛҮҮЛЖ хадгална. Ижил provider давхцвал шинэчилнэ. Танигдсан provider-ууд: google-drive · dropbox · google-meet.",
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string",
                    "maxLength": 32
                  },
                  "access_token": {
                    "type": "string",
                    "maxLength": 4096
                  },
                  "refresh_token": {
                    "type": "string",
                    "maxLength": 4096
                  },
                  "expires_at_ms": {
                    "type": "integer",
                    "description": "epoch мс (0 бол хугацаагүй)"
                  }
                },
                "required": [
                  "provider",
                  "access_token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/integrations/{provider}": {
      "delete": {
        "summary": "Интеграцийг салгах",
        "description": "ИДЕМПОТЕНТ — холбоогүй байсан ч 200.",
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IntegrationProvider"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/me/signature": {
      "get": {
        "summary": "Гарын үсгийн зураг авах",
        "description": "Гарын үсгийн зураг нь Google Drive-д байршдаг; API зөвхөн URL хадгална. Хэрэглэгчийн ID нь ЗӨВХӨН JWT-ээс гардаг тул өөр хүний гарын үсэг рүү хандах боломжгүй.",
        "tags": [
          "assets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "summary": "Гарын үсгийн зураг хадгалах",
        "tags": [
          "assets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AssetURL"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Гарын үсгийн зураг устгах",
        "tags": [
          "assets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/me/orgstamp/{regNo}": {
      "get": {
        "summary": "Байгууллагын тамганы дардас авах",
        "description": "Тухайн байгууллагын АЛЬ НЭГ эрхийн төлөөлөгч байхад хангалттай. Төлөөлөгч биш бол 403 — регистрийн дугаар таамаглаж бусдын тамгыг татахаас (IDOR) хамгаална.",
        "tags": [
          "assets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgRegNo"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "summary": "Байгууллагын тамганы дардас хадгалах (ADMIN)",
        "description": "Зөвхөн ADMIN эрхтэй төлөөлөгч. MANAGER эрхтэй хүн 403 авна.",
        "tags": [
          "assets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgRegNo"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AssetURL"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Байгууллагын тамганы дардас устгах (ADMIN)",
        "tags": [
          "assets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgRegNo"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/oauth2/auth": {
      "get": {
        "summary": "OAuth2 authorization endpoint",
        "description": "`redirect_uri` нь бүртгэлтэйтэй ЯГ (тэмдэгт бүрээр) тулгагдана — prefix/wildcard ХЭЗЭЭ Ч биш. Client эсвэл redirect_uri буруу бол алдааг RP руу ЧИГЛҮҮЛЭХГҮЙ, шууд харуулна (баталгаажаагүй хаяг руу чиглүүлэх боломжгүй). Public client-д PKCE (`S256`) ЗААВАЛ. Амжилттай бол нэвтрэх хуудас руу 302. `/api/v1`-ээс гадуур.",
        "tags": [
          "oidc"
        ],
        "security": [],
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Client ID"
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Бүртгэгдсэн redirect URI"
          },
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "code (өөр утга дэмжигдэхгүй)"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Зайгаар тусгаарласан scope-ууд"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "RP-ийн opaque төлөв"
          },
          {
            "name": "nonce",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "id_token-д давтагдана"
          },
          {
            "name": "code_challenge",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "PKCE challenge (base64url)"
          },
          {
            "name": "code_challenge_method",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Зөвхөн S256"
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "OIDC prompt"
          }
        ],
        "responses": {
          "302": {
            "description": "Нэвтрэх хуудас эсвэл RP руу (алдааны үед)"
          },
          "400": {
            "description": "RFC 6749 §5.2 алдаа",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "error_description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}